Add API to limit system promos
There are some applications that want to limit system promos. This can
be due to various reasons (i.e. app is showing its own promo and showing
another promo will confuse and overwhelm the user).

This API is also specifically needed for the App-to-web feature. The
feature has multiple cases where it is not optimal to switch the user to
the web application; therefore, we do not want to show users the
education promo which urges the user to try switching to the web. The
full description of this use case can be found here:
go/app-to-web-education

Test: m
Bug: 373680430
API-Coverage-Bug: 376903053
Flag:com.android.window.flags.enable_desktop_windowing_app_to_web_education

Change-Id: I7de8dd71ae2609b4b09e2fdf739aa3b4ba67a3e3