Add debug.tracing.profile.* system properties
Adds debug.tracing.profile_boot_classpath and debug.tracing.profile_system_server system properties, so that Perfetto can show when the device is in Boot Classpath or System Server profiling mode. The performance of the system and interpretation of traces is affected by enabling or disabling this profiling.

BUG=372503978

Change-Id: I20400cab34b4f06bde2e75e87010175cff48f58b
Flag: EXEMPT small change, tested