ForegroundService: Include HR, SkinTemp, Spo2 permissions for health fgs.
We are deprecating the BODY_SENSORS permission in favor of
READ_HEART_RATE, READ_SKIN_TEMPERATURE, and READ_OXYGEN_SATURATION.
As part of this we want to ensure that apps using any of those
permissions are able to start a foreground service for the
"health" type.

Based on feedback from the API council, we are marking the API as a
@FlaggedApi which should ensure our changes aren't reflected until
the flag is flipped.

Flag: android.permission.flags.replace_body_sensor_permission_enabled
Test: Built Wear emulator, atest BootCompletedFgsStartTest ActivityManagerForegroundServiceTypeTest
Bug: 379178288
Bug: 379178636
Bug: 375252723
Change-Id: I304731995f5f2d7b5e85dfd267fbca5ff4ab75fe