Introduce VibrationEffect.Composition delay type
Introduce new parameter to Composition.addPrimitive method to specify
the type of delay that should be applied to the vibration primitive.

The current behaviour will remain the default type, which is a pause
between the previous primitive and the new one.

A new delay type is defined as relative offset, representing the time
after the start of the previous primitive that this new one should start
playing. This allows compositions to be defined with regular timings,
independently of the primitive duration on the specific device.

Bug: 373357740
Flag: android.os.vibrator.primitive_composition_absolute_delay
Test: FrameworksVibratorServicesTests
Change-Id: I2d9a9d46318fdea1740af81f2ac45b4f6dc7359d