Remove mActiveAdmin from EnforcingAdmin
The field was only set when EnforcingAdmin was created for authorization
purposes, but not when it was loaded from XML, so the getter can't be
reliably used on all instances. All EnfrocingAdmins created for a DPC or
DA have component name set, which can be used to recover ActiveAdmin if
necessary.

Bug: 335663055
Test: btest a.d.c.WipeDataTest
Test: btest a.d.c.LockTest
Test: btest a.d.c.PasswordComplexityTest
Flag: android.app.admin.flags.active_admin_cleanup
Change-Id: Iabd5a4e9ee4a309ef4e781faaf2900a55004772b