Use correct context, resources and configuration in QS package
Thanks to a linter that highlights non-annotated usages of resources, we realized several classes were not using the correct context.

Here:
- A few classes in the biometrics package were using ShadeDisplayAware context and resources, when they should have used the global one (for the default display), as they were tied to the finger print sensor (that at least for now, is only on the device)
- In a few other classes, @Application context and @Main resources have been used, as they are unrelated to the shade (and are not planning to move between displays)
- Many QS classes were not using the correct context. Now all those classes are either annotated with @ShadeDisplayAware, or @Main/@Application.

Generally, there shouldn't be not-annotated usages of context, resources, Layout inflator or Configuration{Repository|Interactor|State} after this cl.
There are 2 exceptions that are being clarified.

The expectation is that when the flag is off, there are no runtime changes (as the instances provided, even after these changes, will be exactly the same).

Bug: 362719719
Bug: 374267505
Test: presubmits + e2e tests
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Icf802edb383824866ba2ea07c45457d93dfa1591