Delete generated preview file if corrupted
Update AppWidgetServiceImpl so that if it fails to read a generated
preview file, the file is deleted and providers are updated.

This change also updates the service to keep previews after a provider
app updates. Since we now delete previews files if they cannot be read
into RemoteViews, we do not need to worry about missing resource IDs
after an app update.

Bug: 364420494
Test: atest CtsAppWidgetHostTestCases
Flag: android.appwidget.flags.remote_views_proto
Change-Id: I7f4a8aa6c73471c176aa75f2e47627ad8574ce3a