Avoid media updates due to changing constraints
In QQS we want media to take as much space as possible. As the column is
bounded, when the tiles change size due to squishiness, we don't want
the constraints to media change (unbounded) as it will wrap to its size.

Therefore, we pass a requiredHeight max of infinity.

Test: manual, perfetto traces to see that media is not being updated on
expansion
Bug: 383085298
Flag: com.android.systemui.qs_ui_refactor_compose_fragment

Change-Id: Id85fef7b3e38a9f271d549e7142139b76a5161e6